+++ lib/spear/connection.ex
@@ -211,13 +211,14 @@ defmodule Spear.Connection do
   def handle_cast(:connect, s), do: {:connect, s.config, s}
 
   def handle_cast({:push, request_ref, message}, s) when is_reference(request_ref) do
-    # TODO backpressure and http2 window sizes
+    # YARD backpressure and http2 window sizes
     with %Request{rpc: rpc} <- s.requests[request_ref],
          {wire_data, _size} =
            Spear.Request.to_wire_data(message, rpc.service_module, rpc.request_type),
          {:ok, conn} <- Mint.HTTP2.stream_request_body(s.conn, request_ref, wire_data) do
       {:noreply, put_in(s.conn, conn)}
     else
+      # coveralls-ignore-start
       nil ->
         {:noreply, s}
 
@@ -225,6 +226,7 @@ defmodule Spear.Connection do
         s = put_in(s.conn, conn)
 
         if reason == @closed, do: {:disconnect, :closed, s}, else: {:noreply, s}
+        # coveralls-ignore-stop
     end
   end